feat: Synchronization mode, snapshots, multi-queues, yaml config
- feat: recognize yaml output from --config execution
- feat: Synchronization and Event modes for kubernetes hooks
- feat: Synchronization mode return all existing objects in binding context
- feat: Event mode is usual mode with one object
- feat: include filterResult in bindingContext
- feat: includeSnapshots to include existing objects from kubernetes bindings #92
- feat: multi task queues, to execute hooks in parallel #88
- feat: use logrus for json logging
- feat: binding context generator for hook testing by m.nabokikh
- refactor: remove code duplicated in addon-operator
- fix: klog problem #87

+++ test/utils/kind-k8s-cluster.go
@@ -20,12 +20,17 @@ func KindCreateCluster(clusterName string) error {
 		return err
 	}
 	if known {
-		return fmt.Errorf("a cluster with the name %q already exists", clusterName)
+		if os.Getenv("KIND_USE_CLUSTER") == "" {
+			return fmt.Errorf("a cluster with the name '%s' is already exists", clusterName)
+		} else {
+			// Cluster is already exists and user ask to use it.
+			return nil
+		}
 	}
 
 	// create a cluster context and create the cluster
 	ctx := cluster.NewContext(clusterName)
-	fmt.Printf("Creating cluster %q ...\n", clusterName)
+	fmt.Printf("Creating cluster '%s' ...\n", clusterName)
 	if err = ctx.Create(
 		create.WithConfigFile(""),
 		create.WithNodeImage(KindNodeImage()),
@@ -45,8 +50,11 @@ func KindCreateCluster(clusterName string) error {
 }
 
 func KindDeleteCluster(clusterName string) error {
+	if os.Getenv("KIND_USE_CLUSTER") != "" {
+		return nil
+	}
 	// Delete the cluster
-	fmt.Printf("Deleting cluster %q ...\n", clusterName)
+	fmt.Printf("Deleting cluster '%s' ...\n", clusterName)
 	ctx := cluster.NewContext(clusterName)
 	if err := ctx.Delete(); err != nil {
 		return errors.Wrap(err, "failed to delete cluster")
@@ -71,6 +79,10 @@ func KindClusterVersion() string {
 //
 // See: https://hub.docker.com/r/kindest/node/tags
 func KindNodeImage() string {
+	image := os.Getenv("KIND_IMAGE")
+	if image != "" {
+		return image
+	}
 	images := map[string]string{
 		"1.11": "kindest/node:v1.11.10",
 		"1.12": "kindest/node:v1.12.10",
@@ -81,4 +93,32 @@ func KindNodeImage() string {
 	}
 
 	return images[KindClusterVersion()]
+}
+
+func KindClusterName(clusterPrefix string) string {
+	name := os.Getenv("KIND_CLUSTER_NAME")
+	if name != "" {
+		return name
+	}
+
+	forceName := os.Getenv("KIND_USE_CLUSTER")
+	if forceName != "" {
+		return forceName
+	}
+
+	clusterVer := KindClusterVersion()
+	if clusterVer != "" {
+		return fmt.Sprintf("%s-%s", clusterPrefix, clusterVer)
+	} else {
+		return clusterPrefix
+	}
+}
+
+func KindUseClusterMessage(clusterName string) string {
+	forceName := os.Getenv("KIND_USE_CLUSTER")
+	if forceName != "" {
+		return fmt.Sprintf("Use 'kind' flavour of k8s cluster with name '%s'", clusterName)
+	}
+	return fmt.Sprintf("Use 'kind' flavour of k8s cluster v%s with node image %s", KindClusterVersion(), KindNodeImage())
+
 }
